requirements.txt
python-dateutil
requests
spotipy
sqlalchemy
click